requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
mod-wsgi==4.7.1
glob2==0.7
python-dateutil==2.8.1